requirements.txt
git+https://github.com/huggingface/transformers.git
pillow
PyQt6